return list

+++ vue-project/src/views/AddBorrowList.vue
@@ -1,17 +1,27 @@
 <template>
   <div>
-    <p>เพิ่มรายการยืมครุภัณฑ์: {{ code }}</p>
+    <b-field 
+      v-model='lblResult'
+      v-bind:type='isSuccessType'
+      v-bind:label='isSuccessMsg'>
+    </b-field>
+    <!-- <p>เพิ่มรายการยืมครุภัณฑ์: {{ code }}</p> -->
   </div>
 </template>
 <script>
+  import Swal from 'sweetalert2';
   import firestore from '../../../firebase-config/vue/firebase';
-  const line = require('../../../line-config/config')
+
+  const line = require('../../../line-config/config');
+
   export default {
     data() {
       return {
-        photo: "https://www.freeiconspng.com/uploads/no-image-icon-11.PNG",
         code: this.$route.params.code,
-        userProfile: null
+        userProfile: null,
+        lblResult:null,
+        isSuccessType:null,
+        isSuccessMsg:null
       }
     },
     mounted(){
@@ -36,7 +46,7 @@
         console.error('Error initialize LIFF: ', err)
       });
     },
-    created() {
+    created() { // ค้นเลขครุภัณฑ์ในฐานข้อมูล
       const docRef = firestore.collection('items');
         const query = docRef
           .where('item_code','==',this.code)
@@ -44,11 +54,34 @@
         query
         .get()
         .then(snapshot =>{
-          snapshot.forEach((doc) => {
-            console.log(doc.id);
-            localStorage.setItem("item:"+this.code, this.code);
-            liff.closeWindow()
-          });
+          if(!snapshot.empty){ // หากพบข้อมูลและ status = 'ใช้งาน' สามารถยืมได้
+            snapshot.forEach((doc) => {
+              Swal.fire({
+                title: 'เพิ่มรายการสำเร็จ',
+                text: 'เพิ่มรายการยืมครุภัณฑ์: '+ doc.data().item_code +' เรียบร้อยแล้วค่ะ',
+                icon: 'success'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  this.isSuccessType = 'is-success'
+                  this.isSuccessMsg = 'เพิ่มรายการยืมครุภัณฑ์: ' + doc.data().item_code
+                  localStorage.setItem('item:'+this.code, this.code);
+                  liff.closeWindow()
+                }
+              })
+            });
+          }else{ // หากไม่พบข้อมูลหรือ status != 'ใช้งาน' ไม่สามารถยืมได้
+            Swal.fire({
+              title: 'ไม่สามารถเพิ่มข้อมูลได้',
+              text: 'เนื่องจากไม่มีข้อมูลครุภัณฑ์: '+ this.code +' ในระบบหรือไม่สามารถยืมได้ค่ะ',
+              icon: 'error'
+            }).then((result) => {
+              if (result.isConfirmed) {
+                this.isSuccessType = 'is-danger'
+                this.isSuccessMsg = 'ไม่สามารถเพิ่มรายการยืมครุภัณฑ์: ' + this.code +' ได้เนื่องจากไม่มีข้อมูลในระบบหรือไม่สามารถยืมได้ค่ะ'
+                liff.closeWindow()
+              }
+            })
+          }
         })
         .catch(err =>{
           console.log(err);
